+++ app/cdap/components/WranglerGrid/AddTransformation/ColumnTable/__test__/ColumnTable.test.tsx
@@ -14,36 +14,97 @@
  * the License.
  */
 
-import { render } from '@testing-library/react';
+import { fireEvent, render, screen } from '@testing-library/react';
 import React from 'react';
 import ColumnTable from 'components/WranglerGrid/AddTransformation/ColumnTable';
+import T from 'i18n-react';
 
-describe('Should test ColumnTable Component', () => {
-  it('Should render ColumnTable Component', () => {
-    const mockDirectiveFunctionSupportedDataType = ['all'];
-    const container = render(
+describe('It should test FunctionNameWidget Component', () => {
+  it('Should render the FunctionNameWidget Component', () => {
+    render(
       <ColumnTable
         columns={[
-          {
-            name: 'body_0',
-            label: 'body_0',
-            type: ['String'],
-          },
+          { name: 'a', type: ['test'], label: 'test' },
+          { name: 'a', type: ['test'], label: 'test' },
         ]}
-        transformationFunctionSupportedDataType={mockDirectiveFunctionSupportedDataType}
-        onSingleSelection={jest.fn()}
+        transformationDataType={['all']}
+        onSingleSelection={() => jest.fn()}
         selectedColumns={[]}
-        dataQualityValue={[
-          {
-            label: 'body_0',
-            value: '0',
-          },
-        ]}
-        isSingleSelection={true}
-        handleDisableCheckbox={jest.fn()}
-        onMultipleSelection={jest.fn()}
+        dataQualityValue={[]}
+        isSingleSelection={false}
+        handleDisableCheckbox={() => false}
+        onMultipleSelection={() => jest.fn()}
+        totalColumnCount={0}
+        setSelectedColumns={() => jest.fn()}
+        transformationName={'swap-columns'}
       />
     );
-    expect(container).toBeDefined();
+    const checkBoxElement = screen.getByTestId(/column-table-check-box/i);
+    fireEvent.click(checkBoxElement);
+    expect(checkBoxElement).toBeInTheDocument();
+    expect(screen.getByTestId(/column-table-parent/i)).toBeInTheDocument();
+    expect(screen.getByTestId(/panel-columns/i)).toHaveTextContent(
+      `${T.translate('features.WranglerNewUI.GridPage.addTransformationPanel.columns')}`
+    );
+  });
+
+  it('Should render the FunctionNameWidget Component with data type as test', () => {
+    const f = [
+      { name: 'string', label: 'string', type: ['test1', 'mock'] },
+      { name: 'string', label: 'string', type: ['test2', 'mock'] },
+      { name: 'string', label: 'string', type: ['Abhilash', 'IronMan'] },
+    ];
+    render(
+      <ColumnTable
+        columns={[{ name: 'a', type: ['test'], label: 'test' }]}
+        transformationDataType={['test']}
+        onSingleSelection={() => jest.fn()}
+        selectedColumns={f}
+        dataQualityValue={[]}
+        isSingleSelection={false}
+        handleDisableCheckbox={() => false}
+        onMultipleSelection={() => jest.fn()}
+        totalColumnCount={0}
+        setSelectedColumns={() => jest.fn()}
+        transformationName={'swap-columns'}
+      />
+    );
+    const checkBoxElement = screen.getByTestId(/column-table-check-box/i);
+    fireEvent.click(checkBoxElement);
+    expect(screen.getByTestId(/panel-values/i)).toHaveTextContent(
+      `${T.translate('features.WranglerNewUI.GridPage.addTransformationPanel.nullValues')}`
+    );
+    expect(screen.getByTestId(/column-table-parent/i)).toBeInTheDocument();
+  });
+
+  it('Should trigger the checkboxs handleChange when colums length is greater than 2', () => {
+    const f = [
+      { name: 'string', label: 'string', type: ['test1', 'mock'] },
+      { name: 'string', label: 'string', type: ['test2', 'mock'] },
+      { name: 'string', label: 'string', type: ['Abhilash', 'IronMan'] },
+    ];
+    const handleChange = jest.fn();
+    render(
+      <ColumnTable
+        columns={[
+          { name: 'a', type: ['test'], label: 'test' },
+          { name: 'a', type: ['test'], label: 'test' },
+          { name: 'a', type: ['test'], label: 'test' },
+        ]}
+        transformationDataType={['test']}
+        onSingleSelection={() => jest.fn()}
+        selectedColumns={[]}
+        dataQualityValue={[]}
+        isSingleSelection={false}
+        handleDisableCheckbox={() => false}
+        onMultipleSelection={() => jest.fn()}
+        totalColumnCount={0}
+        setSelectedColumns={() => jest.fn()}
+        transformationName={'swap-columns'}
+      />
+    );
+    const checkBoxElement = screen.getByTestId(/column-table-check-box/i);
+    fireEvent.click(checkBoxElement);
+    expect(checkBoxElement).toBeInTheDocument();
   });
 });
